pyproject.toml
[project]
name = "ru-chat-bot"
version = "1.0.0"
description = "An intelligent chatbot."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "optimum[onnxruntime]>=1.11.0",
    "aiohttp>=3.8.0",
    "feedparser>=6.0.0",
    "spacy>=3.5.0",
    "pymorphy3>=1.2.0",
    "dateparser>=1.1.0",
    "python-dotenv>=1.0.0",
    "faiss-cpu>=1.7.0",
    "ru-wiki-search @ git+https://github.com/KvaytG/ru-wiki-search.git",
    "ru-toxicity-detector @ git+https://github.com/KvaytG/ru-toxicity-detector.git"
]

[project.optional-dependencies]
dev = ["pytest", "black"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["chat_bot*"]

[tool.setuptools.package-data]
"chat_bot" = [
    "resources/cities.json",
    "resources/cities-synonyms.json",
    "resources/dssm-answers.json",
    "resources/dssm-index.faiss",
    "resources/model.pt"
]
